py: Extract floating point routines from objfloat.c.
Especially the fp_is* functions are useful to have in the public
API, and will be used in upcoming commits regarding JSON.
